requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
dotnetcore2>=2.1.16
azureml
azureml-dataset-runtime
azureml.train
azureml.core